requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-Login==0.6.2
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
